+++ src/gui/static/src/app/app.module.ts
@@ -46,7 +46,6 @@ import { MatMenuModule } from '@angular/material/menu';
 import { MatProgressBarModule } from '@angular/material/progress-bar';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatSelectModule } from '@angular/material/select';
-import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatTooltipModule } from '@angular/material/tooltip';
@@ -110,6 +109,8 @@ import { HwRemovePinDialogComponent } from './components/layout/hardware-wallet/
 import { HwUpdateFirmwareDialogComponent } from './components/layout/hardware-wallet/hw-update-firmware-dialog/hw-update-firmware-dialog.component';
 import { HwUpdateAlertDialogComponent } from './components/layout/hardware-wallet/hw-update-alert-dialog/hw-update-alert-dialog.component';
 import { ChangeNoteComponent } from './components/pages/send-skycoin/send-preview/transaction-info/change-note/change-note.component';
+import { MsgBarComponent } from './components/layout/msg-bar/msg-bar.component';
+import { MsgBarService } from './services/msg-bar.service';
 
 
 const ROUTES = [
@@ -248,6 +249,7 @@ const ROUTES = [
     HwUpdateFirmwareDialogComponent,
     HwUpdateAlertDialogComponent,
     ChangeNoteComponent,
+    MsgBarComponent,
   ],
   entryComponents: [
     AddDepositAddressComponent,
@@ -297,7 +299,6 @@ const ROUTES = [
     MatProgressBarModule,
     MatProgressSpinnerModule,
     MatSelectModule,
-    MatSnackBarModule,
     MatTabsModule,
     MatToolbarModule,
     MatTooltipModule,
@@ -332,6 +333,7 @@ const ROUTES = [
     HwWalletSeedWordService,
     LanguageService,
     StorageService,
+    MsgBarService,
     DecimalPipe,
   ],
   bootstrap: [AppComponent],