+++ src/gui/static/src/app/components/pages/send-skycoin/send-form-advanced/send-form-advanced.component.ts
@@ -1,10 +1,10 @@
 import { Component, EventEmitter, Input, OnDestroy, OnInit, Output, ViewChild, ChangeDetectorRef } from '@angular/core';
 import { WalletService } from '../../../../services/wallet.service';
 import { FormArray, FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { MatDialog, MatSnackBar, MatDialogConfig } from '@angular/material';
+import { MatDialog, MatDialogConfig } from '@angular/material';
 import { PasswordDialogComponent } from '../../../layout/password-dialog/password-dialog.component';
 import { ButtonComponent } from '../../../layout/button/button.component';
-import { showSnackbarError, getHardwareWalletErrorMsg } from '../../../../utils/errors';
+import { getHardwareWalletErrorMsg } from '../../../../utils/errors';
 import { Subscription, ISubscription } from 'rxjs/Subscription';
 import { NavBarService } from '../../../../services/nav-bar.service';
 import { SelectAddressComponent } from './select-address/select-address';
@@ -22,6 +22,7 @@ import { DoubleButtonActive } from '../../../layout/double-button/double-button.
 import { PriceService } from '../../../../services/price.service';
 import { SendFormComponent } from '../send-form/send-form.component';
 import { ChangeNoteComponent } from '../send-preview/transaction-info/change-note/change-note.component';
+import { MsgBarService } from '../../../../services/msg-bar.service';
 
 @Component({
   selector: 'app-send-form-advanced',
@@ -66,7 +67,7 @@ export class SendFormAdvancedComponent implements OnInit, OnDestroy {
     private appService: AppService,
     private formBuilder: FormBuilder,
     private dialog: MatDialog,
-    private snackbar: MatSnackBar,
+    private msgBarService: MsgBarService,
     private navbarService: NavBarService,
     private hwWalletService: HwWalletService,
     private translate: TranslateService,
@@ -147,7 +148,7 @@ export class SendFormAdvancedComponent implements OnInit, OnDestroy {
     this.closeSyncCheckSubscription();
     this.subscriptions.unsubscribe();
     this.navbarService.hideSwitch();
-    this.snackbar.dismiss();
+    this.msgBarService.hide();
     this.destinationSubscriptions.forEach(s => s.unsubscribe());
   }
 
@@ -240,7 +241,7 @@ export class SendFormAdvancedComponent implements OnInit, OnDestroy {
   }
 
   private prepareTransaction() {
-    this.snackbar.dismiss();
+    this.msgBarService.hide();
     this.previewButton.resetState();
     this.sendButton.resetState();
 
@@ -490,7 +491,7 @@ export class SendFormAdvancedComponent implements OnInit, OnDestroy {
           this.processingSubscription = this.walletService.injectTransaction(transaction.encoded, note)
             .subscribe(noteSaved => {
               if (note && !noteSaved) {
-                showSnackbarError(this.snackbar, this.translate.instant('send.error-saving-note'));
+                this.msgBarService.showError(this.translate.instant('send.error-saving-note'));
               }
 
               this.showSuccess();
@@ -657,8 +658,8 @@ export class SendFormAdvancedComponent implements OnInit, OnDestroy {
 
   private showError(error) {
     this.busy = false;
+    this.msgBarService.showError(error);
     this.navbarService.enableSwitch();
-    showSnackbarError(this.snackbar, error);
     this.previewButton.resetState().setEnabled();
     this.sendButton.resetState().setEnabled();
   }