+++ src/gui/static/src/app/components/pages/wallets/change-name/change-name.component.ts
@@ -7,9 +7,9 @@ import { ButtonComponent } from '../../../layout/button/button.component';
 import { MessageIcons } from '../../../layout/hardware-wallet/hw-message/hw-message.component';
 import { HwWalletService } from '../../../../services/hw-wallet.service';
 import { TranslateService } from '@ngx-translate/core';
-import { MatSnackBar } from '@angular/material';
-import { getHardwareWalletErrorMsg, showSnackbarError } from '../../../../utils/errors';
+import { getHardwareWalletErrorMsg } from '../../../../utils/errors';
 import { ISubscription } from 'rxjs/Subscription';
+import { MsgBarService } from '../../../../services/msg-bar.service';
 
 enum States {
   Initial,
@@ -48,7 +48,7 @@ export class ChangeNameComponent implements OnInit, OnDestroy {
     private walletService: WalletService,
     private hwWalletService: HwWalletService,
     private translateService: TranslateService,
-    private snackbar: MatSnackBar,
+    private msgBarService: MsgBarService,
   ) {}
 
   ngOnInit() {
@@ -70,7 +70,7 @@ export class ChangeNameComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
-    this.snackbar.dismiss();
+    this.msgBarService.hide();
     if (this.hwConnectionSubscription) {
       this.hwConnectionSubscription.unsubscribe();
     }
@@ -88,7 +88,7 @@ export class ChangeNameComponent implements OnInit, OnDestroy {
       return;
     }
 
-    this.snackbar.dismiss();
+    this.msgBarService.hide();
     this.button.setLoading();
 
     this.finishRenaming(this.form.value.label);
@@ -123,7 +123,7 @@ export class ChangeNameComponent implements OnInit, OnDestroy {
               response.errorMsg = getHardwareWalletErrorMsg(this.hwWalletService, this.translateService, err);
               this.dialogRef.close(response);
             } else {
-              showSnackbarError(this.snackbar, getHardwareWalletErrorMsg(this.hwWalletService, this.translateService, err));
+              this.msgBarService.showError(getHardwareWalletErrorMsg(this.hwWalletService, this.translateService, err));
               this.currentState = States.Initial;
               if (this.button) {
                 this.button.resetState();