+++ src/gui/static/src/app/utils/errors.ts
@@ -1,4 +1,3 @@
-import { MatSnackBar, MatSnackBarConfig } from '@angular/material';
 import { HwWalletService, OperationResults } from '../services/hw-wallet.service';
 import { TranslateService } from '@ngx-translate/core';
 import { AppConfig } from '../app.config';
@@ -25,13 +24,6 @@ export function parseResponseMessage(body: string): string {
   }
 
   return body;
-}
-
-export function showSnackbarError(snackbar: MatSnackBar, body: string, duration = 300000) {
-  const config = new MatSnackBarConfig();
-  config.duration = duration;
-
-  snackbar.open(parseResponseMessage(body), null, config);
 }
 
 export function getHardwareWalletErrorMsg(hwWalletService: HwWalletService, translateService: TranslateService, error: any): string {